Improve the initialize function about read in Poppler slide.
Add a new function to dindin.c which named keep_read_slide()
It read one more slide and answer that should we read more slides or not.
Put this function in glutIdleFunc so that Dindin only read slides while idleing.
It means we don't have to wait Dindin read whole slides and can still use it.